Smith College and Miscellaneous Changes
Here, I have added more specific information to the commentaries in the Smith collection to indicate whether they are held/presented by the department or the museum. I have also been working through a stack of miscellaneous folders (in the USEP office) and have created/edited entries along the way.